add missing methods to contract, extract ContractContents into shared class to eliminate code duplication
