+++ src/main/java/adjoint/uplink_sdk/client/parameters/wrappers/ContractWrap.java
@@ -15,6 +15,7 @@
  */
 package adjoint.uplink_sdk.client.parameters.wrappers;
 
+import adjoint.uplink_sdk.client.parameters.wrappers.ContractContents;
 import adjoint.uplink_sdk.client.Response;
 
 
@@ -22,27 +23,10 @@ import adjoint.uplink_sdk.client.Response;
  * @author Adjoint Inc.
  */
 public class ContractWrap extends Response {
-  public final ContractContent contents;
+  public final ContractContents contents;
 
-  public ContractWrap(String tag, ContractContent contents) {
+  public ContractWrap(String tag, ContractContents contents) {
     super(tag, "RPCResp");
     this.contents = contents;
-  }
-
-
-  class ContractContent {
-    public final Integer timestamp;
-    public final String script;
-    public final Storage storage;
-
-    ContractContent(Integer timestamp, String script, Storage storage) {
-      this.timestamp = timestamp;
-      this.script = script;
-      this.storage = storage;
-    }
-  }
-
-  class Storage {
-    // nothing
   }
 }
